feat(MangaFire): Refactor
- Change interceptors.ts to network.ts
- Use new cloudflare bypass detection method
- Extract out functions to utils
- Move all parsers to separate file
- Fix sorting options selector
- Fix rating value is out 10
- Simplify display of status
- Clean up language utilities
- Add missing license headers
- Set languages on home section queries
- Fix Result interface not matching runtime checks
- Remove weird languagesState in settings form
- Respect languages on discover section fetch
- Clean up searchUrl URL builder
