+++ src/MangaFire/forms.ts
@@ -1,3 +1,6 @@
+/* SPDX-License-Identifier: GPL-3.0-or-later */
+/* Copyright © 2026 Inkdex */
+
 import {
   AdvancedSearchForm,
   Form,
@@ -9,42 +12,7 @@ import {
 } from "@paperback/types";
 
 import type { SearchDetails, SearchMetadata } from "./models";
-
-// Language Helper Class
-class MFLanguagesClass {
-  Languages = [
-    { name: "English", MDCode: "en", flagCode: "🇬🇧", default: true },
-    { name: "Français", MDCode: "fr", flagCode: "🇫🇷" },
-    { name: "Español", MDCode: "es", flagCode: "🇪🇸" },
-    { name: "Español (Latinoamérica)", MDCode: "es-la", flagCode: "🇲🇽" },
-    { name: "Português", MDCode: "pt", flagCode: "🇵🇹" },
-    { name: "Português (Brasil)", MDCode: "pt-br", flagCode: "🇧🇷" },
-    { name: "日本語", MDCode: "ja", flagCode: "🇯🇵" },
-  ];
-
-  constructor() {
-    // Sort languages by name
-    this.Languages = this.Languages.sort((a, b) => (a.name > b.name ? 1 : -1));
-  }
-
-  getCodeList(): string[] {
-    return this.Languages.map((language) => language.MDCode);
-  }
-
-  getName(code: string): string {
-    return this.Languages.find((language) => language.MDCode === code)?.name ?? "Unknown";
-  }
-
-  getFlagCode(code: string): string {
-    return this.Languages.find((language) => language.MDCode === code)?.flagCode ?? "🏳️";
-  }
-
-  getDefault(): string[] {
-    return this.Languages.filter((language) => language.default).map((language) => language.MDCode);
-  }
-}
-
-export const MFLanguages = new MFLanguagesClass();
+import { MFLanguages } from "./utils/language";
 
 export function getLanguages(): string[] {
   return (Application.getState("languages") as string[] | undefined) ?? MFLanguages.getDefault();
@@ -214,25 +182,15 @@ export class MangaFireAdvancedSearchForm extends AdvancedSearchForm {
 
 // Main Settings Form
 export class MangaFireSettingsForm extends Form {
-  private languagesState: {
-    value: string[];
-    updateValue: (newValue: string[]) => Promise<void>;
-  };
+  private languages: string[];
 
   constructor() {
     super();
-    const languages = getLanguages();
-    this.languagesState = {
-      value: languages,
-      updateValue: async (newValue: string[]) => {
-        this.languagesState.value = newValue;
-        setLanguages(newValue);
-      },
-    };
+    this.languages = getLanguages();
   }
 
   async updateValue(value: string[]): Promise<void> {
-    this.languagesState.value = value;
+    this.languages = value;
     setLanguages(value);
   }
 
@@ -247,7 +205,7 @@ export class MangaFireSettingsForm extends Form {
           SelectRow("languages", {
             title: "Languages",
             subtitle: (() => {
-              const selectedLangCodes = this.languagesState.value;
+              const selectedLangCodes = this.languages;
               const selectedLangNames = selectedLangCodes
                 .map(
                   (langCode) =>
@@ -256,7 +214,7 @@ export class MangaFireSettingsForm extends Form {
                 .sort();
               return selectedLangNames.join(", ");
             })(),
-            value: this.languagesState.value,
+            value: this.languages,
             options: MFLanguages.getCodeList().map((code) => ({
               id: code,
               title: `${MFLanguages.getFlagCode(code)} ${MFLanguages.getName(code)}`,
