+++ src/MangaFire/network.ts
@@ -1,0 +1,43 @@
+/* SPDX-License-Identifier: GPL-3.0-or-later */
+/* Copyright © 2026 Inkdex */
+
+import {
+  CloudflareError,
+  PaperbackInterceptor,
+  type Request,
+  type Response,
+} from "@paperback/types";
+
+import { DOMAIN } from "./models";
+
+export class MangaFireInterceptor extends PaperbackInterceptor {
+  async interceptRequest(request: Request): Promise<Request> {
+    return {
+      ...request,
+      headers: {
+        ...request.headers,
+        referer: `${DOMAIN}/`,
+        "user-agent": await Application.getDefaultUserAgent(),
+      },
+    };
+  }
+
+  override async interceptResponse(
+    request: Request,
+    response: Response,
+    data: ArrayBuffer,
+  ): Promise<ArrayBuffer> {
+    const cfMitigated = response.headers?.["cf-mitigated"];
+    if (cfMitigated === "challenge") {
+      throw new CloudflareError({
+        url: `${DOMAIN}/`,
+        method: request.method ?? "GET",
+        headers: {
+          "user-agent": await Application.getDefaultUserAgent(),
+        },
+      });
+    }
+
+    return data;
+  }
+}
