+++ src/MangaFire/pbconfig.ts
@@ -1,3 +1,6 @@
+/* SPDX-License-Identifier: GPL-3.0-or-later */
+/* Copyright © 2026 Inkdex */
+
 import { ContentRating, SourceIntents, type ExtensionInfo } from "@paperback/types";
 
 export default {
